// src/app/shared/project/interfaces/project.interface.ts
export interface ProjectInterface {
    logo: string;
    title: string;
    type: string;
    participants?: string;
    dates: string;
    FeTechnologies: string;
    BeTechnologies?: string;
    MobileTechnologies?: string;
    layout?: string;
    difficulty: string;
    description: string;
    adaptivity?: string;
    url?: string;
    role: string;
    stack: string;
    achievements: string;
}